update birdset implementation
now includes both EfficientNetB0 and ConvNext models. Does not depend on birdset repo or datasets package for determining the correct preprocessing. Instead, includes local implementation of preprocessing for each model based on examples in the birdset repo.

Training is still pretty slow, it's unclear why. Not unreasonably slow perhaps: 17 minutes for 1 epoch of 9,000 samples on ConvNext on an apple silicon M1 chip. 12 minutes for EfficientNetB1.