upgrade HawkEars to v1.0.x
we support loading the low-band (Ruffed Grouse/ Spruce Grouse) model or the embedding model instead of the standard ensemble of 6 CNNs.

Versions 1.0.0 - 1.0.8 should all work. HawkEars v2.0 (coming Fall 2025) will likely not be compatible.

to use the HawkEars that was previously available in the model zoo, use the class HawkEars_v010 (version 0.1.0)

resolves Add HawkEarsRuffedGrouse model #26